packagegroup-image-full: add pulseaudio-misc
The much I dislike pulseaudio when jamming, it is necesessary to set up
defaults for pulseaudio (at least in full images): As soon as pulseaudio-server
is installed, alsa's default is pulseaudio.
If more than one audio card (or MIDI device) is connected, it is necessary to
set some defaults for pulseaudio.
To get default soundcards desired's name 'pactl' is useful [1] and it is
shipped by pulseaudio-misc.

[1] https://wiki.archlinux.org/index.php/PulseAudio/Examples#Set_the_default_output_source

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>